requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# AI/LLM
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langsmith>=0.1.0
tiktoken>=0.5.0

# Statistical modeling
statsmodels>=0.14.0
prophet>=1.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0

# Google Sheets
google-api-python-client>=2.108.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.5.0

# Async HTTP
httpx>=0.24.0
aiohttp>=3.9.0
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
pytest-benchmark>=4.0.0
pytest-html>=4.1.0
pytest-timeout>=2.2.0

# Utilities
python-dateutil>=2.8.2
tzdata>=2024.1
PyYAML>=6.0
tenacity>=8.2.0
tqdm>=4.66.0
joblib>=1.3.0
packaging>=23.0
apscheduler>=3.10.0
python-multipart>=0.0.6
orjson>=3.9.0

pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

slowapi>=0.1.9

# Database
DATABASE_URL=sqlite:///./agentic_analyst.db

# JWT
SECRET_KEY=your-secret-key-change-in-production
ALGORITHM=HS256
ACCESS_TOKEN_EXPIRE_MINUTES=30

# Email (Gmail example)
SMTP_HOST=smtp.gmail.com
SMTP_PORT=587
SMTP_USER=your-email@gmail.com
SMTP_PASSWORD=Agentic-Analyst
FROM_EMAIL=alexkavanagh6@gmail.com

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Email
aiosmtplib==2.0.1
jinja2==3.1.2

argon2-cffi==23.1.0
passlib[argon2]==1.7.4
